Chapter 20 pre challenges Extrabeous picture data solved by using camera app, as per Xamarin's guide. Not using the BNR CimreCameraActivity or CrimeCameraFragment anymore, but using BNR's resize and display methods.
* CrimeCameraFragment.cs: No Functional change

* CrimeFragment.cs: Used Camera app instead of BNR's implementation
  with a surface view which had problems with extra info in picture.
Using Xamarin guide's take and get photo methods using Camera app, but
  BNR's resizing and displaying code.

* CriminalIntent.csproj: Don't compile CrimeCameraActivity and
  Fragment

* PictureUtils.cs: Type fixes

* Resource.designer.cs: blah

+++ BNR_Android_Book/CriminalIntent/CriminalIntent/PictureUtils.cs
@@ -25,18 +25,18 @@ namespace CriminalIntent
 			float srcWidth = options.OutWidth;
 			float srcHeight = options.OutHeight;
 
-			int inSampleSize = 1;
+			double inSampleSize = 1.0;
 			if (srcHeight > destHeight || srcWidth > destWidth) {
 				if (srcWidth > srcHeight) {
-					inSampleSize = (int)Math.Round(srcHeight / destHeight);
+					inSampleSize = Math.Round(srcHeight / destHeight);
 				}
 				else {
-					inSampleSize = (int)Math.Round(srcWidth / destWidth);
+					inSampleSize = Math.Round(srcWidth / destWidth);
 				}
 			}
 
 			options = new BitmapFactory.Options();
-			options.InSampleSize = inSampleSize;
+			options.InSampleSize = (int)inSampleSize;
 
 			Bitmap bitmap = BitmapFactory.DecodeFile(path, options);
 			return new BitmapDrawable(a.Resources, bitmap);
